feat: Implement infinit indexing of chunks
This uses cantor pairing in conjunction with mapping Z -> N. The only
issue is that I can actually not use this haha, because the pattern in
which it indexes the levels is too complex for me to adjust the levels
to. So I will probably just focus on N^2 -> N which makes the pattern
much simpler.