feat!: Implement proper item y-sorting
This is achieved by using `YSort` and properly offsetting the value
(taking into account the LDtk layer index and the parent z transform).

Note that overlapping tiles in the same layer is not supported yet. The
reason is simple: If you have a tile in layer A, which has an index of
3, then this tile will have a z value of 3 in bevy. However, if you have
two tiles that are overlapping each other, one of them will have value 3
while the other will have value 4.

That's why we are currently restricted to only non-overlapping tiles
(which is totally fine for this project).

More info at https://github.com/Trouv/bevy_ecs_ldtk/issues/276.