feat!: Finish enemy spawner system (mostly)
It works good enough for now. There are most likely going to be tweaks
here and there, but the foundation is done now. This was much harder
then expected, but the end result is quite okay I think. Still room for
making it simpler but it will suffice for now.